requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
cytoolz==0.12.2
dask==2023.5.0
einops==0.7.0
filelock==3.13.1
fire==0.5.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
grpcio==1.60.0
idna==3.6
imagecodecs-lite==2019.12.3
imageio==2.31.5
importlib-metadata==7.0.0
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
locket==1.0.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1
packaging==23.2
partd==1.4.1
Pillow==10.1.0
pip==23.3.1
ply==3.11
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.8.2
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.0
termcolor==2.3.0
tifffile==2020.6.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
unicodedata2==15.1.0
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.42.0
win-inet-pton==1.1.0
yarl==1.9.3
zipp==3.17.0
